+++ animation.h
@@ -6,15 +6,15 @@
 typedef struct {
 	SDL_Rect *frames;      /* frames array */
 	size_t length;         /* number of frames in animation */
-	unsigned int reset;    /* index to reset to upon reaching the end of the frames array */
-	unsigned int interval; /* interval of index incrementation (for automatic animations)*/
+	size_t reset;          /* index to reset to upon reaching the end of the frames array */
+	Uint32 interval;       /* interval of index incrementation (for automatic animations)*/
 } Animation;
 
 /* Create an animation, length must be equal to the size of the frames array */
-Animation *CreateAnimation( size_t length, SDL_Rect* frames, unsigned int reset, unsigned int interval );
+Animation *CreateAnimation( size_t length, SDL_Rect* frames, size_t reset, Uint32 interval );
 
 /* Initialize an existing animation, length must be equal to the size of the frames array */
-Animation *InitAnimation( Animation *anim, size_t length, SDL_Rect* frames, unsigned int reset, unsigned int interval );
+Animation *InitAnimation( Animation *anim, size_t length, SDL_Rect* frames, size_t reset, Uint32 interval );
 
 /* free the animation frames array */
 void FreeAnimation( Animation *anim );
@@ -23,7 +23,7 @@ void FreeAnimation( Animation *anim );
 SDL_Rect *GetNextFrame( Animation *anim, size_t *index );
 
 /* get the updated frame in the animation based on the time difference */
-SDL_Rect *GetUpdatedFrame( Animation *anim, size_t *index, unsigned int *time );
+SDL_Rect *GetUpdatedFrame( Animation *anim, size_t *index, Uint32 *time );
 
 #endif /* _ANIMATION_H_ */
